Refactor login authentication tests to TypeScript with improved structure
